+++ src/Type/UrlType.php
@@ -6,17 +6,11 @@
 namespace Zicht\Bundle\UrlBundle\Type;
 
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Form\FormInterface;
-use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Zicht\Bundle\AdminBundle\Form\AutocompleteType;
 use Zicht\Bundle\UrlBundle\Aliasing\Aliasing;
-use Zicht\Bundle\UrlBundle\Form\DataTransformer\TextTransformer;
 
-/**
- * Type for choosing an URL
- */
 class UrlType extends AbstractType
 {
     /**
@@ -29,11 +23,6 @@ class UrlType extends AbstractType
         $this->aliasing = $aliasing;
     }
 
-    public function getBlockPrefix()
-    {
-        return 'zicht_url';
-    }
-
     public function configureOptions(OptionsResolver $resolver)
     {
         parent::configureOptions($resolver);
@@ -41,40 +30,24 @@ class UrlType extends AbstractType
         $resolver
             ->setDefaults(
                 [
-                    'with_edit_button' => true,
-                    'no_transform_public' => false,
-                    'no_transform_internal' => false,
-                    'url_suggest' => '/admin/url/suggest',
+                    'repo' => 'Zicht\Bundle\UrlBundle\Entity\UrlAlias',
+                    'transformer' => 'noop', // disable all transformation. we do this ourselves with UrlTypeAutocompleteDataTransformer
                 ]
             );
     }
 
-    public function finishView(FormView $view, FormInterface $form, array $options)
-    {
-        parent::finishView($view, $form, $options);
-        $view->vars['url_suggest'] = $options['url_suggest'];
-        $view->vars['with_edit_button'] = $options['with_edit_button'];
-    }
-
     public function getParent()
     {
-        return TextType::class;
+        return AutocompleteType::class;
+    }
+
+    public function getBlockPrefix()
+    {
+        return 'zicht_url';
     }
 
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $mode = TextTransformer::MODE_TO_INTERNAL | TextTransformer::MODE_TO_PUBLIC;
-
-        if ($options['no_transform_public']) {
-            $mode ^= TextTransformer::MODE_TO_PUBLIC;
-        }
-
-        if ($options['no_transform_internal']) {
-            $mode ^= TextTransformer::MODE_TO_INTERNAL;
-        }
-
-        if ($mode > 0) {
-            $builder->addModelTransformer(new TextTransformer($this->aliasing, $mode));
-        }
+        $builder->addViewTransformer(new UrlTypeAutocompleteDataTransformer($this->aliasing));
     }
 }